// src/GridCartographer/schema.ts
export interface Entry {
  /** The index number of the palette entry. */
  i: number;

  /** The color value of the entry expressed in HTML notation #RRGGBB. */
  rgb: string;

  /** If set to 1 this indicates a color that has been edited. */
  edit?: "1";
}

export interface Export {
  /** The friendly version number of Grid Cartographer used to export the document. This takes the form year.month.suffix. */
  from: string;

  /** The date of export in YYYY-MM-DD format. */
  date: string;

  /** The (local) time of export in HH:MM:SS format. */
  time: string;
}

export interface Floor {
  /** The number of the floor. Negative values are basements, zero is the ground floor and positive values are the floors above. */
  index: number;

  tiles: {
    bounds: Bounds;
    rows?: Row[];
  };

  notes: Note[];
}

export interface Map {
  export: Export;
  regions: Region[];
  palette?: Entry[];
}

export interface Note {
  /** The X co-ordinate of the note. */
  x: number;

  /** The Y co-ordinate of the note given in the co-ordinate space specified by the origin attribute of the setup element. */
  y: number;

  __data?: string;
}

export interface Region {
  /** The floor count of this region. Excluding ground floor if disabled. */
  floor_count: number;

  /** The index of the lowest floor. Negative values are basements. */
  lowest_floor: number;

  /** The shape of the grid used for all floors in this region. */
  grid_shape: "square" | "hexh" | "hexv";

  /** The name of a region or tilemap. */
  name: string;

  setup: {
    /** Either tl or bl which, respectively, specify either a top-left or bottom-left grid origin. */
    origin: "tl" | "bl";
  };

  floors: Floor[];
}