// src/GridCartographer/enums.ts
export const enum Edge {
  None = 0,
  Wall = 1,
  Door = 2,
  Door_Locked = 3,
  Door_Hidden = 4,
  Door_OneWayLU = 5,
  Door_Hidden_OneWayLU = 6,
  Wall_OneWayLU = 7,
  Door_OneWayRD = 8,
  Door_Hidden_OneWayRD = 9,
  Wall_OneWayRD = 10,
  Door_Empty = 12,
  Wall_Secret = 13,
  Door_Trapped = 14,
  Door_HalfL = 15,
  Door_HalfR = 16,
  Wall_HalfL = 17,
  Wall_HalfR = 18,
  ButtonLU = 19,
  ButtonRD = 20,
  TorchLU = 21,
  TorchRD = 22,
  LeverLU = 23,
  LeverRD = 24,
  Bars = 25,
  TorchDouble = 26,
  Gate = 27,
  Message = 28,
  Door_Secret = 29,
  NicheLU = 30,
  NicheRD = 31,
  Keyhole = 32,
  Door_Box = 33,
  Wall_Trapped = 34,
}

export const enum InBlockEdge {
  None = 0,
  Horizontal = 1,
  Vertical = 2,
  DiagonalLeft = 3,
  DiagonalRight = 4,
  CornerNW = 5,
  CornerNE = 6,
  CornerSE = 7,
  CornerSW = 8,
}

export const enum Marker {
  None = 0,
  StairsUp = 1,
  StairsDown = 2,
  NPC = 3,
  TeleportIn = 4,
  TeleportOut = 5,
  RotatingRoom = 6,
  Pit_Open = 7,
  Death = 8,
  Entrance = 9,
  Exit = 10,
  Turntable = 11,
  Chest_Open = 12,
  Key = 13,
  Monster = 14,
  Switch = 15,
  Fountain = 16,
  SavePoint = 17,
  Target = 18,
  Pressure = 19,
  Pentagram = 20,
  Elevator = 21,
  Zap = 22,
  Unknown = 23,
  Event = 24,
  Message = 25,
  LadderUp = 26,
  LadderDown = 27,
  StairsFacingNorth = 34,
  StairsFacingEast = 35,
  StairsFacingSouth = 36,
  StairsFacingWest = 37,
  Wardrobe = 38,
  Coin_Single = 39,
  Bookshelf = 40,
  ShapeCross = 41,
  LadderTwoWays = 44,
  Chest_Closed = 46,
  Chest_Trapped = 47,
  Chest_Locked = 48,
  Ore = 49,
  Pit_Covered = 50,
  Pit_Trapped = 51,
  Well = 52,
  ShapeTriangle = 53,
  ShapeSmallSquare = 54,
  ShapeSquare = 55,
  ShapeSmallCircle = 56,
  ShapeCircle = 57,
  GemDiamond = 58,
  GemEmerald = 59,
  GemRuby = 60,
  GemCrystal = 61,
  ArrowUp = 62,
  ArrowRight = 63,
  ArrowDown = 64,
  ArrowLeft = 65,
  Sack = 66,
  Map = 67,
  Purse = 68,
  Barrel = 69,
  RampUp = 70,
  RampDown = 71,
  Boulder = 72,
  Stone = 73,
  PressurePlateAnd_Stone = 74,
  ArrowLR = 78,
  ArrowUD = 79,
  ArrowDiagUL = 80,
  ArrowDiagUR = 81,
  ArrowDiagDR = 82,
  ArrowDiagDL = 83,
  DoubleArrowULDR = 84,
  DoubleArrowURDL = 85,
  Vial = 86,
  Coffin = 87,
  CoffinWithCruciform = 88,
  Tree = 97,
  Shop = 98,
  Bed = 99,
  Tavern = 100,
  Health = 101,
  MoveableBlock = 102,
  TrainerOrTeacher = 103,
  Skull = 104,
  Bones = 105,
  Boat = 106,
  Bridge = 107,
  Signpost = 108,
  Pillar = 109,
  Armor = 110,
  Grave = 111,
  Statue = 112,
  ArrowUL = 117,
  ArrowUR = 118,
  ArrowDL = 119,
  ArrowDR = 120,
  ArrowLU = 121,
  ArrowRU = 122,
  ArrowLD = 123,
  ArrowRD = 124,
  Weapons = 125,
  Boots = 126,
  Altar = 127,
  Food = 128,
  Scroll = 130,
  Book = 131,
  HarvestPlant = 132,
  TimberPile = 133,
  Doorway = 134,
  Tent = 135,
  Spring = 136,
}

export const enum Terrain {
  None = 0,
  PencilRegularA = 1,
  PencilRegularB = 2,
  PencilRegularC = 3,
  PencilRegularD = 4,
  PencilRegularE = 5,
  PencilRegularF = 6,
  PencilRegularG = 7,
  PencilRoughA = 10,
  PencilRoughB = 11,
  PencilRoughC = 12,
  PencilRoughD = 13,
  PencilRoughE = 14,
  PencilRoughF = 15,
  PencilDebrisA = 20,
  PencilDebrisB = 21,
  PencilDebrisC = 22,
  PencilDebrisD = 23,
  Inside = 31,
  Outside = 32,
  Water = 33,
  Lava = 34,
  Rock = 35,
  Vegetation = 36,
  Ooze = 38,
  Block = 39,
  Sand = 40,
  Wood = 41,
  PencilCrackedA = 50,
  PencilCrackedB = 51,
  PencilCrackedC = 52,
  PencilCrackedD = 53,
  PencilCrackedE = 54,
  PencilCrackedF = 55,
  PencilCrackedG = 56,
  PencilCrackedH = 57,
  PencilCrackedI = 58,
  PencilCrackedJ = 59,
  Metal = 75,
  Trees = 76,
  Snow = 95,
  Mountain = 113,
  Track = 115,
}